Moved file upload event handler to a module
